requirements.txt
hid==1.0.4
Pillow==7.1.1
PyEventEmitter==1.0.5
